% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_distribution.R
\name{Exact_distribution}
\alias{Exact_distribution}
\title{Exact Distribution for WSNR}
\usage{
Exact_distribution(n, alpha, side = "two.sided")
}
\description{
This function is hidden and not intended for direct use.
}
\keyword{internal}
