% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits_asymptotical.R
\name{Limits_asymptotical}
\alias{Limits_asymptotical}
\title{Limits Asymptotical}
\usage{
Limits_asymptotical(n, alpha, side = "two.sided")
}
\description{
This function is hidden and not intended for direct use.
}
\keyword{internal}
