% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chart.R
\name{plot_chart}
\alias{plot_chart}
\title{Plot Control Chart}
\usage{
plot_chart(ic, statistics, ucl, lcl, name = NULL)
}
\arguments{
\item{ic}{The value for the process center or target.}

\item{statistics}{A numeric vector of statistics to be plotted.}

\item{ucl}{The upper control limit.}

\item{lcl}{The lower control limit.}

\item{name}{A character string specifying the name of the statistics (default: NULL).}
}
\description{
Plots a control chart with control limits and statistics.
}
\examples{
statistics <- c(1.2, 1.5, 1.4, 1.7, 2.0)
ucl <- 2.5
lcl <- 1.0
plot_chart(ic = 1.8, statistics = statistics, ucl = ucl, lcl = lcl, name = "Data")
}
