% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shewhart_snr.R
\name{shewhart_snr}
\alias{shewhart_snr}
\title{Shewhart Chart using Wilcoxon Signed-Rank Test}
\usage{
shewhart_snr(
  X,
  mu0 = NULL,
  alpha = 0.0027,
  group_by_col = FALSE,
  exact = FALSE,
  side = "two.sided"
)
}
\arguments{
\item{X}{A matrix or data frame of numeric values.}

\item{mu0}{the process median, if known.
If NULL, it will be estimated using the mean of the observations.}

\item{alpha}{The significance level for the control limits (default: 0.0027).}

\item{group_by_col}{Logical indicating whether to treat columns as different groups (default: FALSE).}

\item{exact}{Logical indicating whether to use the exact distribution for control limits calculation (default: FALSE).}

\item{side}{A character string specifying the alternative hypothesis for the test ("two.sided", "greater", or "less") (default: "two.sided").}
}
\description{
Calculates the Shewhart Chart using the Wilcoxon Signed-Rank Test.
}
\examples{
X <- matrix(rnorm(100), nrow = 10)
shewhart_snr(X, mu0 = 0, alpha = 0.01, group_by_col = TRUE)
}
